backend/requirements.txt
fastapi
uvicorn
pdfminer.six
python-docx
scikit-learn
python-multipart
